requirements.txt
python-telegram-bot==20.6
google-generativeai==0.5.0
aiohttp
